services/vector-store/main.py
"""
Vector Store Service - Manages conversation memory and document embeddings
for context retrieval and RAG-based verification.
"""

import os
from contextlib import asynccontextmanager
from typing import AsyncGenerator

import structlog
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel

load_dotenv()

logger = structlog.get_logger()


class Document(BaseModel):
    """A document to store in the vector database."""

    id: str
    content: str
    metadata: dict = {}
    source: str
    source_url: str | None = None


class SearchQuery(BaseModel):
    """Query for semantic search."""

    query: str
    top_k: int = 5
    filter_metadata: dict | None = None


class SearchResult(BaseModel):
    """A single search result."""

    id: str
    content: str
    score: float
    metadata: dict
    source: str


class SearchResponse(BaseModel):
    """Response containing search results."""

    results: list[SearchResult]
    query: str


class ConversationMemory(BaseModel):
    """Store conversation context."""

    conversation_id: str
    messages: list[dict]


@asynccontextmanager
async def lifespan(app: FastAPI) -> AsyncGenerator[None, None]:
    """Application lifespan manager."""
    logger.info("vector_store_service_starting")
    # TODO: Initialize database connection
    # TODO: Create tables if not exist
    yield
    logger.info("vector_store_service_stopping")


app = FastAPI(
    title="Jarvis Vector Store Service",
    description="Manages embeddings and semantic search for conversation memory",
    version="0.1.0",
    lifespan=lifespan,
)


@app.get("/health")
async def health_check() -> dict:
    """Health check endpoint."""
    return {"status": "ok", "service": "vector-store"}


@app.post("/documents")
async def store_document(document: Document) -> dict:
    """
    Store a document with its embedding in the vector database.
    """
    logger.info(
        "store_document",
        doc_id=document.id,
        source=document.source,
        content_length=len(document.content),
    )

    # TODO: Implement actual storage
    # - Generate embedding using OpenAI or local model
    # - Store in PostgreSQL with pgvector

    return {"stored": True, "id": document.id}


@app.post("/search", response_model=SearchResponse)
async def search_documents(query: SearchQuery) -> SearchResponse:
    """
    Perform semantic search across stored documents.
    """
    logger.info("search_query", query=query.query, top_k=query.top_k)

    # TODO: Implement actual search
    # - Generate query embedding
    # - Perform cosine similarity search
    # - Return top_k results

    return SearchResponse(results=[], query=query.query)


@app.post("/conversations/{conversation_id}/memory")
async def store_conversation_memory(
    conversation_id: str, memory: ConversationMemory
) -> dict:
    """
    Store conversation context for retrieval.
    """
    logger.info(
        "store_conversation",
        conversation_id=conversation_id,
        message_count=len(memory.messages),
    )

    # TODO: Implement conversation memory storage
    # - Store messages with embeddings
    # - Enable retrieval of relevant context

    return {"stored": True, "conversation_id": conversation_id}


@app.get("/conversations/{conversation_id}/context")
async def get_conversation_context(
    conversation_id: str, query: str | None = None, limit: int = 10
) -> dict:
    """
    Retrieve relevant conversation context.
    """
    logger.info(
        "get_context", conversation_id=conversation_id, query=query, limit=limit
    )

    # TODO: Implement context retrieval
    # - If query provided, do semantic search
    # - Otherwise return recent messages

    return {"conversation_id": conversation_id, "context": [], "query": query}


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", "8004")),
        reload=os.getenv("ENV", "development") == "development",
    )
