services/vector-store/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
openai==1.9.0
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
numpy==1.26.3
python-dotenv==1.0.0
structlog==24.1.0
